cream_modal/pyproject.toml
[tool.poetry]
name = "cream-modal"
version = "0.1.0"
description = ""
authors = ["Dominique Paul <dominique.c.a.paul@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
websockets = "^15.0.1"
logfire = "^3.11.0"
torch = "^2.6.0"
modal = ">=0.73.134,<0.74.0"
diffusers = ">=0.32.2,<0.33.0"
transformers = ">=4.50.3,<5.0.0"
sentencepiece = ">=0.2.0,<0.3.0"
accelerate = ">=1.5.2,<2.0.0"
jupyter = ">=1.1.1,<2.0.0"
safetensors = ">=0.5.3,<0.6.0"
pillow = ">=11.1.0,<12.0.0"
numpy = ">=2.2.4,<3.0.0"
matplotlib = ">=3.10.1,<4.0.0"
fastapi = ">=0.115.12,<0.116.0"
xformers = "^0.0.29.post3"

[tool.poetry.group.dev.dependencies]
ipython = "^9.0.2"
ipykernel = "^6.29.5"
iprogress = "^0.4"
ipywidgets = ">=8.1.5,<9.0.0"
fire = "^0.7.0"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
